refactoring analyzer code
